added prepublishOnly script to add new version to changelog on non-beta publishes
